+++ prov/kubernetes/k8s_utils_test.go
@@ -15,10 +15,12 @@
 package kubernetes
 
 import (
-	"context"
+	"strings"
 	"testing"
 
+	appsv1 "k8s.io/api/apps/v1beta1"
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/api/extensions/v1beta1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -115,36 +117,219 @@ func TestGetNoneExternalIPAdress(t *testing.T) {
 	}
 }
 
-func TestWaitForPVCDeletionAndDeleted(t *testing.T) {
-	k8s := newTestK8s()
-	errorChan := make(chan struct{})
-	finishedChan := make(chan struct{})
-	getCount := 0
-	ctx := context.Background()
+/* Return a map of mocked k8s objects corresponding to their resource in
+Currently support only : StatefulSet, PVC, Service and Deployment */
+func mockSupportedResources() map[string]runtime.Object {
 	pvc := &corev1.PersistentVolumeClaim{ObjectMeta: metav1.ObjectMeta{Name: "pvcTest", Namespace: "test-ns"}}
-	//Simulate a deletion in progress : wait for 2 get that return the volume and then fakely delete it
-	//If the API continue to receive GET, raise error
-	k8s.clientset.(*fake.Clientset).Fake.AddReactor("get", "persistentvolumeclaims", func(action k8stesting.Action) (bool, runtime.Object, error) {
+	dep := &v1beta1.Deployment{ObjectMeta: metav1.ObjectMeta{Name: "deploymentTest", Namespace: "test-ns"}}
+	sts := &appsv1.StatefulSet{ObjectMeta: metav1.ObjectMeta{Name: "statefulSetTest", Namespace: "test-ns"},
+		Spec: appsv1.StatefulSetSpec{}}
+	svc := &corev1.Service{ObjectMeta: metav1.ObjectMeta{Name: "serviceTest", Namespace: "test-ns"}}
+	supportedRes := make(map[string]runtime.Object)
+	supportedRes["persistentvolumeclaims"] = pvc
+	supportedRes["deployments"] = dep
+	supportedRes["statefulsets"] = sts
+	supportedRes["services"] = svc
+	return supportedRes
+}
+
+/*  React to get on k8s objects. After 2 get if deleteObj boolean is set to true then it fakely delete it. If the boolean is set to false,
+then it marked the object as succefully present or deployed for pods controllers.
+If the API continue to receive GET, raise error by signaling the errorChan */
+func fakeGetObjectReaction(k8sObject runtime.Object, errorChan chan struct{}, deleteObj bool) k8stesting.ReactionFunc {
+	getCount := 0
+	return func(action k8stesting.Action) (bool, runtime.Object, error) {
+		// TODO: manage error another way
+		if action.GetVerb() != "get" {
+			close(errorChan)
+		}
+
 		getCount++
 		if getCount > 5 {
 			close(errorChan)
 		} else if getCount > 2 {
+			if deleteObj {
+				return true, nil, apierrors.NewNotFound(action.GetResource().GroupResource(), action.GetResource().Resource)
+			}
+			return true, k8sObject, nil
+
+		}
+		return true, nil, nil
+	}
+}
+
+func fakeObjectCompletion(k8sObject yorcK8sObject, errorChan chan struct{}) k8stesting.ReactionFunc {
+	getCount := 0
+	return func(action k8stesting.Action) (bool, runtime.Object, error) {
+		if action.GetVerb() != "get" {
+			close(errorChan)
+		}
+		getCount++
+		if getCount > 5 {
+			close(errorChan)
+		} else if getCount > 2 {
+			//Mark obj deployed
+
+			switch obj := k8sObject.(type) {
+			case *yorcK8sDeployment:
+				obj.Status.AvailableReplicas = *obj.Spec.Replicas
+				return true, obj.getObjectRuntime(), nil
+			case *yorcK8sService:
+				return true, obj.getObjectRuntime(), nil
+			case *yorcK8sPersistentVolumeClaim:
+				obj.Status.Phase = corev1.ClaimBound
+				return true, obj.getObjectRuntime(), nil
+			case *yorcK8sStatefulSet:
+				obj.Status.ReadyReplicas = *obj.Spec.Replicas
+				return true, obj.getObjectRuntime(), nil
+			default:
+				close(errorChan)
+			}
+			return true, nil, nil
+		}
+		return true, nil, nil
+	}
+}
+
+func fakeObjectScale(k8sObject yorcK8sObject, errorChan chan struct{}) k8stesting.ReactionFunc {
+	getCount := 0
+	return func(action k8stesting.Action) (bool, runtime.Object, error) {
+		if action.GetVerb() != "get" {
+			close(errorChan)
+		}
+		getCount++
+		if getCount > 5 {
+			close(errorChan)
+		} else if getCount > 2 {
+			//Mark obj deployed
+
+			switch obj := k8sObject.(type) {
+			case *yorcK8sDeployment:
+				obj.Status.AvailableReplicas = *obj.Spec.Replicas
+				return true, obj.getObjectRuntime(), nil
+			case *yorcK8sStatefulSet:
+				obj.Status.ReadyReplicas = *obj.Spec.Replicas
+				return true, obj.getObjectRuntime(), nil
+			default:
+				close(errorChan)
+			}
+			return true, nil, nil
+		}
+		return true, nil, nil
+	}
+}
+
+func fakeObjectDeletion(k8sObject yorcK8sObject, errorChan chan struct{}) k8stesting.ReactionFunc {
+	getCount := 0
+	return func(action k8stesting.Action) (bool, runtime.Object, error) {
+		if action.GetVerb() != "get" {
+			close(errorChan)
+		}
+		getCount++
+		if getCount > 5 {
+			close(errorChan)
+		} else if getCount > 1 {
+			//Mark obj removed
 			return true, nil, apierrors.NewNotFound(action.GetResource().GroupResource(), action.GetResource().Resource)
+			//return true, nil, nil
 		}
-		return true, pvc, nil
-	})
-	go func() {
-		err := waitForPVCDeletion(ctx, k8s.clientset, pvc)
-		if err != nil {
-			t.Logf("Error : %s", err.Error())
-			t.Fatal("Deleted pvc should not raise an error")
-		}
-		close(finishedChan)
-	}()
-	select {
-	case <-errorChan:
-		t.Fatal("Function waitForPVCDeletion is still polling API even though it is deleted")
-	case <-finishedChan:
-		//Wait for test to be well done
+		return true, nil, nil
+	}
+}
+
+func Test_getK8sResourceNamespace(t *testing.T) {
+	depID := "DepID"
+	nsName := "my-namespace"
+	type args struct {
+		deploymentID string
+		k8sResource  yorcK8sObject
+	}
+	tests := []struct {
+		name      string
+		args      args
+		namespace string
+		provided  bool
+	}{
+		{
+			"Test no ns provided",
+			args{depID, &yorcK8sDeployment{}},
+			strings.ToLower(depID),
+			false,
+		},
+		{
+			"Test ns provided",
+			args{depID, &yorcK8sService{ObjectMeta: metav1.ObjectMeta{Name: "pvcTest", Namespace: nsName}}},
+			nsName,
+			true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, got1 := getNamespace(tt.args.deploymentID, tt.args.k8sResource.getObjectMeta())
+			if got != tt.namespace {
+				t.Errorf("getK8sResourceNamespace() namespace = %v, want %v", got, tt.namespace)
+			}
+			if got1 != tt.provided {
+				t.Errorf("getK8sResourceNamespace() provided = %v, want %v", got1, tt.provided)
+			}
+		})
+	}
+}
+
+func Test_podControllersInNamespace(t *testing.T) {
+	nsName := "my-test-namespace"
+	type args struct {
+		namespace string
+	}
+	tests := []struct {
+		name    string
+		setup   func() kubernetes.Interface
+		args    args
+		want    int
+		wantErr bool
+	}{
+		{
+			"Test no controller left",
+			func() kubernetes.Interface {
+				k8s := newTestSimpleK8s()
+				return k8s.clientset
+			},
+			args{namespace: nsName},
+			0, false,
+		},
+		{
+			"Test one deployment left",
+			func() kubernetes.Interface {
+				k8s := newTestSimpleK8s()
+				k8s.clientset.ExtensionsV1beta1().Deployments(nsName).Create(&v1beta1.Deployment{ObjectMeta: metav1.ObjectMeta{Name: "my-deployment", Namespace: nsName}})
+				return k8s.clientset
+			},
+			args{namespace: nsName},
+			1, false,
+		},
+		{
+			"Test one deployment & 2 statefulSets left",
+			func() kubernetes.Interface {
+				k8s := newTestSimpleK8s()
+				k8s.clientset.AppsV1beta1().StatefulSets(nsName).Create(&appsv1.StatefulSet{ObjectMeta: metav1.ObjectMeta{Name: "my-statefulset-1", Namespace: nsName}})
+				k8s.clientset.AppsV1beta1().StatefulSets(nsName).Create(&appsv1.StatefulSet{ObjectMeta: metav1.ObjectMeta{Name: "my-statefulset-2", Namespace: nsName}})
+				k8s.clientset.ExtensionsV1beta1().Deployments(nsName).Create(&v1beta1.Deployment{ObjectMeta: metav1.ObjectMeta{Name: "my-deployment", Namespace: nsName}})
+				return k8s.clientset
+			},
+			args{namespace: nsName},
+			3, false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := podControllersInNamespace(tt.setup(), tt.args.namespace)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("podControllersInNamespace() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if got != tt.want {
+				t.Errorf("podControllersInNamespace() = %v, want %v", got, tt.want)
+			}
+		})
 	}
 }